*: overhaul inter-worker communication
The coupling between workers was a mess with different ways to
ensure some call made it to some worker and the reply got back
to the caller: sometimes an intermediate worker would pick up
one message, send another type and then translate the reply,
other times we had dumb forwarding. The UI thread did a lot of
coordination, which added delays, so we tried to add message
ports for workers to signal each other but these didn't scale:
every worker needed a port for every other worker.

Instead we now have a coordinator worker who drives everything and
routes messages based on a generic envelope format of to and from.
All workers are made by this worker so they can simply postMessage to
to it with a well addressed envelope and let it route.

Many messes were uncovered while we made this change so this refactor
pulls in some additional changes as we. Oh well.

This all started in an attempt to make the zipWorker support remote
zips, so there is also some (commented out) code related to that.